Added Checkmate, scoring
The game will now check for checkmate and has an endstate. Checkmate does not yet include the possibility of dropping a piece in the moves it checks to get out of checkmate. 
The game keeps track of who has played games, and how many times you have won or lost.